refactor:  change srpc_stub_utils name and dir and ajust import references
